add BaseViewPage (#2140)
update ViewPage,MyViewsPage
